added the responsive contact section
