+++ src/app/(home)/page.tsx
@@ -8,6 +8,8 @@ import Values from "@/components/features/values/Values";
 import Footer from "@/components/footer/Footer";
 import HomeNavbar from "@/components/navigation/HomeNavbar";
 import Hero from "./_components/Hero";
+import ContactInfo from "@/components/features/contact/ContactInfo";
+import ContactForm from "@/components/features/contact/ContactForm";
 
 export default function Home() {
   return (
@@ -19,7 +21,13 @@ export default function Home() {
       <AboutUs />
       <MarqueeText />
       <Achievements />
-      <Contact />
+      <Contact>
+        <ContactForm
+          containerStyles={"bg-form-background"}
+          formContainerStyles="px-6 py-16 md:px-8 md:py-24 lg:px-14 lg:py-28 2xl:px-18"
+        />
+        <ContactInfo containerStyles={"md:pl-4 lg:pl-10 xl:pl-18"} />
+      </Contact>
       <MarqueePartners />
       <Footer />
     </main>
